Spin geocode_api off into a separate module
